+++ examples/csv_writing.cpp
@@ -29,4 +29,5 @@ int main()
   csv_writer.append_row(13212123, "AAPL", 100, 210.32321, "BUY");
   csv_writer.append_row(132121123, "META", 300, 478.32321, "SELL");
   csv_writer.append_row(13212123, "AAPL", 120, 210.42321, "BUY");
+  csv_writer.close();
 }