+++ include/quill/sinks/StreamSink.h
@@ -98,49 +98,7 @@ public:
     }
     else
     {
-      // first attempt to create any non-existing directories
-      std::error_code ec;
-      fs::path parent_path;
-
-      if (!_filename.parent_path().empty())
-      {
-        parent_path = _filename.parent_path();
-
-        // The call to fs::status is necessary due to a known issue in GCC versions 8.3.0 to 9.4.0.
-        // In these versions, fs::create_directories(path, ec) internally uses
-        // fs::symlink_status(path, ec) instead of fs::status(path, ec) for checking the path.
-        // This causes a problem because fs::symlink_status does not follow the symlink to the
-        // target directory. As a result,  it fails the is_directory() check but still indicates
-        // that the path exists, leading to a not_a_directory exception being set in the error code
-        auto const st = fs::status(parent_path, ec);
-        if (!is_directory(st))
-        {
-          fs::create_directories(parent_path, ec);
-          if (ec)
-          {
-            // use .string() to also support experimental fs
-            QUILL_THROW(QuillError{std::string{"cannot create directories for path "} +
-                                   parent_path.string() + std::string{" - error: "} + ec.message()});
-          }
-        }
-      }
-      else
-      {
-        parent_path = fs::current_path();
-      }
-
-      // convert the parent path to an absolute path
-      fs::path const canonical_path = fs::canonical(parent_path, ec);
-
-      if (ec)
-      {
-        // use .string() to also support experimental fs
-        QUILL_THROW(QuillError{std::string{"cannot create canonical path for path "} +
-                               parent_path.string() + std::string{" - error: "} + ec.message()});
-      }
-
-      // finally replace the given filename's parent_path with the equivalent canonical path
-      _filename = canonical_path / _filename.filename();
+      _filename = detail::normalize_file_sink_path(_filename);
     }
   }
 
@@ -278,6 +236,17 @@ public:
   }
 
 protected:
+  QUILL_NODISCARD virtual size_t estimate_write_size(
+    MacroMetadata const* /* log_metadata */, uint64_t /* log_timestamp */,
+    std::string_view /* thread_id */, std::string_view /* thread_name */,
+    std::string const& /* process_id */, std::string_view /* logger_name */, LogLevel /* log_level */,
+    std::string_view /* log_level_description */, std::string_view /* log_level_short_code */,
+    std::vector<std::pair<std::string, std::string>> const* /* named_args */,
+    std::string_view /* log_message */, std::string_view log_statement)
+  {
+    return log_statement.size();
+  }
+
   /**
    * Flushes the stream
    */