+++ test/integration_tests/CsvWritingCustomFrontendTest.cpp
@@ -37,6 +37,7 @@ TEST_CASE("csv_writing_custom_frontend")
     csv_writter.append_row(13212123, "AAPL", 100, 210.32321, "BUY");
     csv_writter.append_row(132121123, "META", 300, 478.32321, "SELL");
     csv_writter.append_row(14212123, "AAPL", 120, 210.42321, "BUY");
+    csv_writter.close();
   }
 
   // Wait until the backend thread stops for test stability